fix: major adjustments
