+++ package.json
@@ -20,7 +20,6 @@
     "antd": "^5.3.1",
     "axios": "^1.3.4",
     "jwt-decode": "^3.1.2",
-    "antd": "^5.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.2",
