+++ src/components/common/RegisterInput/Styles.js
@@ -11,12 +11,9 @@ export const Container = styled.div`
 `;
 
 export const Label = styled.label`
-  font-size: 0.8em;
+  font-size: 1em;
   font-family: ${(props) => props.theme.fonts.montserrat};
   font-weight: 400;
-  @media (min-width: 801px) and (max-width: 1030px) {
-    font-size: 0.7em;
-  }
 `;
 
 export const Input = styled.input`
@@ -26,7 +23,7 @@ export const Input = styled.input`
   background-color: transparent;
   padding: 0 5%;
   font-family: ${(props) => props.theme.fonts.montserrat};
-  font-size: 0.8em;
+  font-size: 1em;
   font-weight: 600;
   margin-top: 0.4em;
   border: ${(props) => (props.error ? '2px red solid' : '1px black solid')};
