--- src/components/features/AdminAddCategoryMobile/Styles.js
@@ -1,154 +1,0 @@
-import styled from 'styled-components';
-import { Link } from 'react-router-dom';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  width: 100%;
-  margin-bottom: 50px;
-  align-items: center;
-`;
-
-export const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 400px;
-`;
-
-export const Title = styled.p`
-  font-family: 'Montserrat';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 24px;
-  padding-bottom: 20px;
-
-  color: black;
-`;
-
-export const Subtitle = styled.p`
-  font-family: 'Montserrat';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 24px;
-  padding-bottom: 10px;
-
-  color: black;
-`;
-
-export const SmallInput = styled.input`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  padding: 8px 11px;
-  gap: 10px;
-  margin-bottom: 15px;
-
-  max-width: 100%;
-  width: 100%;
-  height: 36px;
-
-  border: 1px solid black;
-  border-radius: 4px;
-
-  font-family: 'Montserrat';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-`;
-
-export const ButtonsDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const SaveButton = styled.button`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  align-self: center;
-  justify-content: center;
-  padding: 10px;
-  gap: 10px;
-  margin-bottom: 10px;
-
-  width: auto;
-  height: 45px;
-
-  border: 1px solid black;
-  border-radius: 5px;
-  background-color: transparent;
-
-  p {
-    font-family: 'Montserrat';
-    font-style: normal;
-    font-weight: 600;
-    font-size: 18px;
-
-    display: flex;
-    align-items: center;
-    text-align: center;
-
-    color: black;
-
-    :hover {
-      transition: 1s;
-      color: white;
-    }
-  }
-  :hover {
-    transition: 1s;
-    background-color: ${(props) => props.theme.colors.blue};
-    color: white;
-    border: white;
-    cursor: pointer;
-  }
-`;
-
-export const CancelButton = styled(Link)`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  align-self: center;
-  padding: 5px 10px;
-  gap: 10px;
-  text-decoration: none;
-  width: 84px;
-  height: 27px;
-
-  border: 1px solid #000000;
-  border-radius: 5px;
-  background-color: transparent;
-
-  p {
-    font-family: 'Montserrat';
-    font-style: normal;
-    font-weight: 600;
-    font-size: 14px;
-
-    display: flex;
-    align-items: center;
-    text-align: center;
-
-    color: black;
-
-    :hover {
-      transition: 1s;
-      color: white;
-    }
-  }
-
-  :hover {
-    transition: 1s;
-    background-color: red;
-    color: white;
-    border: white;
-    cursor: pointer;
-  }
-`;
