+++ src/components/features/Footer/Styles.js
@@ -1,13 +1,15 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
+const mobileBreakPoint = '850px';
+
 export const Container = styled.div`
   display: flex;
   flex-direction: row;
   height: 300px;
   width: 100%;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     flex-direction: column;
   }
 `;
@@ -21,7 +23,7 @@ export const SideContainer = styled.div`
   align-items: center;
   padding: 0% 7%;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     width: 100%;
     height: auto;
     padding: 3% 7%;
@@ -29,7 +31,7 @@ export const SideContainer = styled.div`
 `;
 
 export const LogoSection = styled.div`
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     padding-top: 30px;
   }
 `;
@@ -44,7 +46,7 @@ export const Text = styled.p`
   color: black;
   padding-top: 16px;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     font-size: 14px;
   }
 `;
@@ -58,7 +60,7 @@ export const MiddleContainer = styled.div`
   align-items: center;
   padding: 0% 7%;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     display: none;
   }
 `;
@@ -72,7 +74,7 @@ export const Tittle = styled.p`
 
   color: #203699;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     font-size: 16px;
   }
 `;
@@ -87,7 +89,7 @@ export const SocialMedias = styled.div`
   margin-bottom: 10px;
   margin-top: 15px;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     margin-bottom: 20px;
     margin-top: 20px;
   }
@@ -135,7 +137,7 @@ export const ButtonMobile = styled.div`
   font-size: 15px;
   margin-bottom: 40px;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     display: flex;
     height: auto;
     justify-content: center;
@@ -151,8 +153,9 @@ export const SectionGoTo = styled.section`
   align-items: center;
   justify-content: center;
 
-  @media (max-width: 850px) {
+  @media (max-width: ${mobileBreakPoint}) {
     padding-bottom: 20px;
+    /* display: none; */
   }
 `;
