+++ src/components/features/ModalEditCategory/ModalEditCategory.jsx
@@ -1,10 +1,12 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import React from 'react';
-import { FiSave } from 'react-icons/fi';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { useState } from 'react';
+
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useQueryClient } from '@tanstack/react-query';
+import PropTypes from 'prop-types';
+import { useForm } from 'react-hook-form';
+import { FiSave } from 'react-icons/fi';
+
+import { useUpdateCategory } from '../../../hooks/query/categories';
 import {
   Container,
   Label,
@@ -13,50 +15,72 @@ import {
   ModalButton,
   Form,
 } from './Styles';
-import { useUpdateCategory } from '../../../hooks/query/categories';
+import {
+  buildUpdateCategoryErrorMessage,
+  updateCategoryValidationSchema,
+} from './utils';
 
-const validationSchema = z.object({
-  name: z
-    .string()
-    .min(1, 'Favor inserir o nome da categoria')
-    .min(3, 'O nome da categoria deve ter pelo menos 3 caracteres')
-    .max(40, 'O nome da categoria deve ter no máximo 40 caracteres'),
-  description: z.string().optional(),
-});
-
-export default function ModalEditCategory({ categoryId, categoryName, close }) {
-  const { handleSubmit, register, watch } = useForm({
-    resolver: zodResolver(validationSchema),
-  });
+export default function ModalEditCategory({ category, close }) {
+  const [isPending, setIsPending] = useState(false); // Important for modal loading
 
   const queryClient = useQueryClient();
-  const onSuccess = () => {
-    queryClient.invalidateQueries({ queryKey: ['categories'] });
-    close();
+  const { mutate: updateCategory } = useUpdateCategory({
+    onSuccess: () => {
+      Promise.all([
+        queryClient.invalidateQueries({
+          queryKey: ['categories', 'searchByName'],
+        }),
+        queryClient.invalidateQueries({
+          queryKey: ['category'],
+        }),
+      ]);
+      close();
+    },
+    onError: (err) => {
+      const errorMessage = buildUpdateCategoryErrorMessage(err);
+
+      // Do something to the errorMessage
+      alert(errorMessage);
+      setIsPending(false);
+    },
+  });
+
+  const {
+    handleSubmit,
+    register,
+    formState: { errors },
+  } = useForm({
+    resolver: zodResolver(updateCategoryValidationSchema),
+  });
+  const onSubmit = (data) => {
+    updateCategory({ _id: category?._id, newCategoryData: data });
+    setIsPending(true);
   };
 
-  const { mutate: updateCategory } = useUpdateCategory({ onSuccess });
-  const onSubmit = (data) =>
-    updateCategory({ id: categoryId, newCategoryData: data });
-  console.log(watch());
   return (
     <Container>
       <ModalContent>
         <Form onSubmit={handleSubmit(onSubmit)}>
           <Label>Nome da categoria:</Label>
           <Input
-            placeholder={categoryName}
+            defaultValue={category?.name}
             id="name"
             name="name"
             type="name"
             {...register('name')}
           />
-          <ModalButton type="submit">
+          {errors?.name?.message && <p>{errors?.name?.message}</p>}{' '}
+          <ModalButton disabled={isPending} type="submit">
             <FiSave size={25} />
-            <p>Salvar Alterações</p>
+            <p>{isPending ? 'Carregando...' : 'Salvar Alterações'}</p>
           </ModalButton>
         </Form>
       </ModalContent>
     </Container>
   );
 }
+
+ModalEditCategory.propTypes = {
+  category: PropTypes.object.isRequired,
+  close: PropTypes.func.isRequired,
+};
