+++ src/components/features/ModalEditCategory/Styles.js
@@ -23,7 +23,7 @@ export const ModalContent = styled.div`
 `;
 
 export const Label = styled.label`
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 500;
   font-size: 24px;
@@ -42,7 +42,7 @@ export const Input = styled.input`
   border: none;
   color: black;
 
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 400;
   font-size: 20px;
@@ -77,7 +77,7 @@ export const ModalButton = styled.button`
   color: white;
 
   p {
-    font-family: 'Montserrat';
+    font-family: ${(props) => props.theme.fonts.montserrat};
     font-style: normal;
     font-weight: 600;
     font-size: 18px;
