+++ src/components/features/ModalEditProduct/Styles.js
@@ -60,7 +60,7 @@ export const CategorySubsection = styled.div`
 `;
 
 export const Text = styled.p`
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 500;
   font-size: 24px;
@@ -82,7 +82,7 @@ export const AddButton = styled.button`
   border: none;
 
   color: white;
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 400;
   font-size: 16px;
@@ -94,7 +94,7 @@ export const AddButton = styled.button`
 `;
 
 export const MiniText = styled.p`
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 500;
   font-size: 16px;
@@ -110,7 +110,7 @@ export const InputModalName = styled.input`
   background: #ffffff;
   border-radius: 4px;
 
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 400;
   font-size: 20px;
@@ -128,7 +128,7 @@ export const InputModal = styled.textarea`
   background: #ffffff;
   border-radius: 4px;
 
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 400;
   font-size: 20px;
@@ -154,7 +154,7 @@ export const ModalButton = styled.button`
   top: 703px;
   padding: 10px;
 
-  font-family: 'Montserrat';
+  font-family: ${(props) => props.theme.fonts.montserrat};
   font-style: normal;
   font-weight: 500;
   font-size: 18px;
