+++ src/components/features/index.js
@@ -4,16 +4,10 @@ export { default as Card } from './Card/Card';
 export { default as Carousel } from './Carousel/Carousel';
 export { default as FormsContactUs } from './FormsContact/FormsContact';
 export { default as AdminMenu } from './AdminMenu/AdminMenu';
-export { default as AdminListCategory } from './AdminListCategory/AdminListCategory';
-export { default as AdminListProduct } from './AdminListProduct/AdminListProduct';
-export { default as ModalAddProduct } from './ModalAddProduct/ModalAddProduct';
-export { default as ModalEditProduct } from './ModalEditCategory/ModalEditCategory';
-export { default as ModalCreateCategory } from './ModalAddProduct/ModalAddProduct';
+export { default as ModalCreateProduct } from './ModalCreateProduct/ModalCreateProduct';
+export { default as ModalEditProduct } from './ModalEditProduct/ModalEditProduct';
+export { default as ModalCreateCategory } from './ModalCreateProduct/ModalCreateProduct';
 export { default as ModalEditCategory } from './ModalEditCategory/ModalEditCategory';
-export { default as AdminEditProductMobile } from './AdminEditProductMobile/AdminEditProductMobile';
-export { default as AdminAddProductMobile } from './AdminAddProductMobile/AdminAddProductMobile';
-export { default as AdminEditCategoryMobile } from './AdminEditCategoryMobile/AdminEditCategoryMobile';
-export { default as AdminAddCategoryMobile } from './AdminAddCategoryMobile/AdminAddCategoryMobile';
 export { default as BudgetForm } from './BudgetForm/BudgetForm';
 export { default as FormsContact } from './FormsContact/FormsContact';
 export { default as FilesList } from './FilesList/FilesList';
