+++ src/layouts/AdminStoreLayout/AdminStoreLayout.jsx
@@ -1,8 +1,9 @@
 import { Outlet } from 'react-router-dom';
+
 import { AdminMenu } from '../../components/features';
 import { Container } from './Styles';
 
-export default function AdminLayout() {
+export default function AdminStoreLayout() {
   return (
     <Container>
       <AdminMenu />
