+++ src/layouts/index.js
@@ -1,2 +1,2 @@
 export { default as AppLayout } from './AppLayout/AppLayout';
-export { default as AdminLayout } from './AdminLayout/AdminLayout';
+export { default as AdminStoreLayout } from './AdminStoreLayout/AdminStoreLayout';
