--- src/pages/AddProduct/AddProduct.jsx
@@ -1,10 +1,0 @@
-import { AdminAddProductMobile } from '../../components/features';
-import { Container } from './Styles';
-
-export default function AddProduct() {
-  return (
-    <Container>
-      <AdminAddProductMobile />
-    </Container>
-  );
-}
