+++ src/pages/CreateCategory/Styles.js
@@ -1,13 +1,123 @@
-/* eslint-disable import/prefer-default-export */
-/* eslint-disable import/prefer-default-export */
+import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 export const Container = styled.div`
   display: flex;
+  align-items: center;
   flex-direction: column;
-  width: 75%;
+  gap: 20px;
 
-  @media (max-width: 700px) {
-    width: 95%;
+  width: 95%;
+  margin-bottom: 50px;
+`;
+
+export const Form = styled.form`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 400px;
+`;
+
+export const Title = styled.p`
+  font-family: ${(props) => props.theme.fonts.montserrat};
+  font-style: normal;
+  font-weight: 700;
+  font-size: 1.5em;
+  line-height: 24px;
+  padding-bottom: 20px;
+
+  color: black;
+`;
+
+export const ButtonsDiv = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const SaveButton = styled.button`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  align-self: center;
+  justify-content: center;
+  padding: 10px;
+  gap: 10px;
+  margin-bottom: 10px;
+
+  width: auto;
+  height: 45px;
+
+  border: 1px solid black;
+  border-radius: 5px;
+  background-color: transparent;
+
+  p {
+    font-family: ${(props) => props.theme.fonts.montserrat};
+    font-style: normal;
+    font-weight: 600;
+    font-size: 18px;
+
+    display: flex;
+    align-items: center;
+    text-align: center;
+
+    color: black;
+
+    :hover {
+      transition: 1s;
+      color: white;
+    }
+  }
+  :hover {
+    transition: 1s;
+    background-color: ${(props) => props.theme.colors.blue};
+    color: white;
+    border: white;
+    cursor: pointer;
+  }
+`;
+
+export const CancelButton = styled(Link)`
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  align-self: center;
+  padding: 5px 10px;
+  gap: 10px;
+  text-decoration: none;
+  width: 84px;
+  height: 27px;
+
+  border: 1px solid #000000;
+  border-radius: 5px;
+  background-color: transparent;
+
+  p {
+    font-family: ${(props) => props.theme.fonts.montserrat};
+    font-style: normal;
+    font-weight: 600;
+    font-size: 14px;
+
+    display: flex;
+    align-items: center;
+    text-align: center;
+
+    color: black;
+
+    :hover {
+      transition: 1s;
+      color: white;
+    }
+  }
+
+  :hover {
+    transition: 1s;
+    background-color: red;
+    color: white;
+    border: white;
+    cursor: pointer;
   }
 `;
