+++ src/services/api/privateEndpoints/endpoints.js
@@ -6,14 +6,14 @@ export const getUsers = async (filters = {}) => {
   return data;
 };
 
-export const updateUser = async ({ id, newUserData }) => {
-  const { data } = await privateApi.put(`/users/${id}`, newUserData);
+export const updateUser = async ({ _id, newUserData }) => {
+  const { data } = await privateApi.put(`/users/${_id}`, newUserData);
 
   return data;
 };
 
-export const deleteUser = async (id) => {
-  const { data } = await privateApi.delete(`/users/${id}`);
+export const deleteUser = async (_id) => {
+  const { data } = await privateApi.delete(`/users/${_id}`);
 
   return data;
 };
@@ -30,8 +30,8 @@ export const createProduct = async (newProduct) => {
   return data;
 };
 
-export const updateCategory = async ({ id, newCategoryData }) => {
-  const data = await privateApi.put(`/categories/${id}`, newCategoryData);
+export const updateCategory = async ({ _id, newCategoryData }) => {
+  const data = await privateApi.put(`/categories/${_id}`, newCategoryData);
 
   return data;
 };
